Add prettier button borders
Add a framed button border style and associated properties

* border_style: 0,1 or 2 for none, shadow, frame;     default 1
* border_colouring: 0,1 for manual, auto;             default 0
* border_grad_dir: 0..6 for down, left, up, right,top-left to bot-right
    or bot-right to top-left; default: 0
* border_grad_top: colour (see note)
* border_grad_bot: colour (see note)

note: border colours set here are only used if border_colouring is
      manual and border_style is frame.

A border_style of frame defines a fixed frame around the button
so that when the button is pressed, the frame stays and the button
has the appearance of moving inward.

A shadow border is a simple border that uses the theme to define
its colour.

Automatic colouring makes the frame border follow the button
colour with a gradient fill from a lighter tint to a darker
shade in the associated direction.