Add a basic scrollable list
Set property 'items' to a semicolon-delimited list for short lists.
Set property 'items_file' to a filename to load the list from
  a file. In this case, a linefeed will be used as the delimiter.
Can select items but no properties are set for the selection.
Deselect items with a second click.

Warning: This widget currently crashes the editor if an item is
pressed in edit mode.